requirements.txt
# Core Libraries
playwright==1.48.0

# Configuration
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-playwright==0.5.2
pytest-mock==3.14.0

# Code Quality (optional but recommended)
# ruff==0.7.4
# mypy==1.13.0
# black==24.10.0